requirements.txt
openai-whisper>=20231117
torch>=2.0.0